emotical_backend/requirements.txt
fastapi==0.111.0
uvicorn==0.28.0
sqlalchemy>=2.0.35
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
requests==2.31.0
aiosqlite==0.20.0
email-validator>=2.1.1
pymysql==1.1.0
pytest==8.1.1
httpx==0.27.0
